cambio cuenta de tiempo

+++ includes/philo.h
@@ -35,7 +35,7 @@ long	time_to_milis(struct timeval t_start, struct timeval t_end);
 void	*ft_calloc(size_t count, size_t size);
 size_t	ft_strlen(const char *s);
 void	milisleep(int max, t_table *table);
-void	status_time(t_philo *philo, t_table *table, char *status, int time);
+void	status_time(t_philo *philo, t_table *table, char *status, size_t time);
 ///Philo
 void	*thread_philo(void *data);
 void	*sniffer_philo(void *table);
